Last Update

+++ Monopoly_Game/src/Monopoly_Game/Chance.java
@@ -5,7 +5,7 @@ import java.awt.*;
 import java.awt.event.*;
 
 public class Chance extends Field{
-    public Chance(String n,double x,double y,String PathImage) {
+    public Chance(String n,int x,int y,String PathImage) {
         super(n,x,y,PathImage);
         
     }
