+++ Monopoly_Game/src/Monopoly_Game/FreeParking.java
@@ -5,7 +5,7 @@ import java.awt.*;
 import java.awt.event.*;
 
 public class FreeParking extends Field{
-    public FreeParking(String n,double x,double y,String PathImage) {
+    public FreeParking(String n,int x,int y,String PathImage) {
         super(n,x,y,PathImage);
         
     }
