+++ Monopoly_Game/src/Monopoly_Game/Go.java
@@ -5,7 +5,7 @@ import java.awt.*;
 import java.awt.event.*;
 
 public class Go extends Field{
-    public Go(String n,double x,double y,String PathImage) {
+    public Go(String n,int x,int y,String PathImage) {
         super(n,x,y,PathImage);
         
     }
