+++ Monopoly_Game/src/Monopoly_Game/GoToJail.java
@@ -5,7 +5,7 @@ import java.awt.*;
 import java.awt.event.*;
 
 public class GoToJail extends Field{
- public GoToJail(String n,double x,double y,String PathImage) {
+ public GoToJail(String n,int x,int y,String PathImage) {
         super(n,x,y,PathImage);
         
     }
