+++ Monopoly_Game/src/Monopoly_Game/Jail.java
@@ -6,7 +6,7 @@ import java.awt.event.*;
 
 public class Jail extends Field{
     
-    public Jail(String n,double x,double y,String PathImage) {
+    public Jail(String n,int x,int y,String PathImage) {
         super(n,x,y,PathImage);
         
     }
